write a relevant message
